backend/core/face.py
import grpc
import psutil
import platform
import GPUtil

class ACEBridge:
    """
    Ponte Facial Local (Inspirada em NVIDIA Audio2Face gRPC)
    Soberania Total: Sem NVIDIA NIM, sem chaves de nuvem.
    """
    def __init__(self, grpc_url="localhost:50051"):
        self.grpc_url = grpc_url
        self.channel = None
        self.stub = None

    async def animate(self, audio_path: str):
        """
        Envia o áudio gerado pelo Piper para o Audio2Face via gRPC local.
        O Audio2Face transforma o áudio em blendshapes em tempo real.
        """
        print(f" [FACE] Enviando áudio {audio_path} para o Audio2Face...")
        # Lógica gRPC real para o A2F
        # stub.PushAudioStream(audio_data)
        pass
