requirements.txt
# requirements.txt para Vaelindra Core (Local)
# Instale com: pip install -r requirements.txt

fastapi==0.110.0
uvicorn==0.27.1
google-generativeai==0.4.0
ollama==0.1.7
faster-whisper==1.0.1
piper-tts==1.2.0
python-dotenv==1.0.1
websockets==12.0
numpy==1.26.4
sounddevice==0.4.6
psutil==5.9.8
GPUtil==1.4.0
obsws-python==1.6.1
pyautogui==0.9.54
aiohttp==3.9.3
Pillow==10.2.0
opencv-python==4.9.0.80
PyYAML==6.0.1
grpcio==1.62.1
grpcio-tools==1.62.1
